move is valid wp cli command to its own hook with new logic
